enhance(analytics): sendBeacon check before script load (#24)
* enhance(analytics): sendBeacon check before load
This introduces a check to the list of checks in the gate before the Fathom
server-side script loads, to check if navigator.sendBeacon exists.

If it does, the script continues as normal. If not, the script short-circuits.

This check ensures that navigator.sendBeacon, a method used by the
server-side Fathom script, is present before the script attempts
to execute it.  If it does and fails, the page the user is visiting
will 404.

* chore: Add changeset

* chore: Add comments for clarity

* fix: check window.navigator, not navigator
?

* Separate checks

* fix: get specific about sendBeacon falsiness

fix: get specific about sendBeacon falsiness...again

* chore: comment clarity

* chore: logic test for canary

* chore: logic test for canary

* chore: prep for merge