Version Packages (#87)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.

# Releases
## @hashicorp/platform-analytics@0.6.1

### Patch Changes

-   [#86](https://github.com/hashicorp/web-platform-packages/pull/86) [`4bf17bb`](https://github.com/hashicorp/web-platform-packages/commit/4bf17bb91ed711eadc60d219eef7a3d9df623d77) Thanks [@nandereck](https://github.com/nandereck)! - Add useErrorPageAnalytics hook