Version Packages (#89)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.

# Releases
## @hashicorp/platform-cli@2.3.0

### Minor Changes

-   [#88](https://github.com/hashicorp/web-platform-packages/pull/88) [`9ad8877`](https://github.com/hashicorp/web-platform-packages/commit/9ad88779846281c2f808d1b2eb54868a8dbf1f07) Thanks [@nandereck](https://github.com/nandereck)! - Uses useErrorPageAnalytics hook for 404 tracking